tosca-esp32c3: Implement device scheme
The implementation of the `tosca` device scheme removes the hard-coded
light device in favor of a more general device.
This increases flexibility in the framework, as there is no longer a
need to maintain or implement specific hard-coded devices.

Device construction is now centered around the `build` function, which
validates developer-provided data against the device scheme.
The result of this process is a verified device that can be passed
to the server.

New errors have been added for devices with missing or invalid mandatory
routes.
