--- crates/tosca-esp32c3/src/devices/mod.rs
@@ -1,2 +1,0 @@
-/// A `light` device.
-pub mod light;
